Make it a separate class
