+++ server/src/main/java/org/elasticsearch/action/search/SearchRequest.java
@@ -31,6 +31,7 @@ import org.elasticsearch.common.io.stream.StreamOutput;
 import org.elasticsearch.common.unit.TimeValue;
 import org.elasticsearch.common.xcontent.ToXContent;
 import org.elasticsearch.search.Scroll;
+import org.elasticsearch.search.builder.PointInTimeBuilder;
 import org.elasticsearch.search.builder.SearchSourceBuilder;
 import org.elasticsearch.search.internal.SearchContext;
 import org.elasticsearch.tasks.TaskId;
@@ -434,7 +435,7 @@ public class SearchRequest extends ActionRequest implements IndicesRequest.Repla
         return source;
     }
 
-    public SearchSourceBuilder.PointInTimeBuilder pointInTimeBuilder() {
+    public PointInTimeBuilder pointInTimeBuilder() {
         if (source != null) {
             return source.pointInTimeBuilder();
         }
